Move alpha tests in their own module
